requirements.txt
mysql-connector-python==9.6.0
pygame==2.6.1
python-nmap==0.7.1
